Refactor: Extract shared box generation code and clean up factory
Major refactoring to improve code quality and eliminate duplication:

1. **Created box_generation.rs module** - Extracted duplicated code from
   renderer.rs and api.rs into a shared module:
   - `generate_box_tree()` - Converts StyledNode to BoxTree
   - `generate_box_node()` - Recursive box generation
   - `is_replaced_element()` - Check for img/video/canvas/etc
   - `create_replaced_box()` - Create replaced element boxes
   - `extract_css()` - Extract CSS from style tags in DOM

2. **Simplified FormattingContextFactory** in factory.rs:
   - Removed unnecessary `_placeholder: ()` field (now unit struct)
   - Removed redundant `create_specific()` method that just delegated to `create()`
   - Updated engine.rs to use `create()` directly

3. **Created golden images directory** - Added tests/fixtures/golden/
   with README placeholder for future visual regression testing

This reduces code duplication by ~200 lines and improves maintainability
by having a single source of truth for box generation logic.
