+++ tests/fixtures/golden/README.md
@@ -1,0 +1,1 @@
+# Golden Reference Images\n\nThis directory will contain golden reference images for visual regression testing.\n\nTo generate golden images, run:\n```bash\nUPDATE_GOLDEN=1 cargo test fixtures\n```
